minor refactor for localizable
