+++ src/fluent.rs
@@ -4,29 +4,6 @@ use log::{error, warn};
 use std::{borrow::Cow, collections::HashMap, error::Error, fs, path::PathBuf};
 use unic_langid::LanguageIdentifier;
 
-pub trait Localizable {
-    fn localize_default(&mut self, _localization: &str) -> &mut Self;
-    fn localize(&mut self, locale: &str, localization: &str) -> &mut Self;
-}
-impl Localizable for serenity::builder::CreateApplicationCommandOption {
-    fn localize_default(&mut self, localization: &str) -> &mut Self {
-        self.description(localization)
-    }
-
-    fn localize(&mut self, locale: &str, localization: &str) -> &mut Self {
-        self.description_localized(locale, localization)
-    }
-}
-impl Localizable for serenity::builder::CreateApplicationCommand {
-    fn localize_default(&mut self, localization: &str) -> &mut Self {
-        self.description(localization)
-    }
-
-    fn localize(&mut self, locale: &str, localization: &str) -> &mut Self {
-        self.description_localized(locale, localization)
-    }
-}
-
 pub struct FluentLocalizer {
     pub fallback_locale: LanguageIdentifier,
     resources: HashMap<LanguageIdentifier, FluentBundle<FluentResource, IntlLangMemoizer>>,
